requirements.txt
discord-rewrite
pyyaml
yarl==0.13.0
oauth2client
gspread
pytz
